Add CLIght, a tool to control house lighting
I finally caved and picked up a Philips Hue for my new office. Being
able to smoothly fade the room's lights to an easy-on-the-eyes cool blue
for late night hacking sessions, or to synchronize the ambient color
temperature with redshift -- all from the comfort of my shell -- feels
incredibly futuristic. Now if I could just integrate an STT engine...