clight: add support for groups and scenes
CLIght is a bit more flexible now! It can apply a list of presets to
multiple lights simultaneously. This mechanism is called "scenes" in
every other Hue application I've seen (and indeed the official API), so
we'll use the same term.

CLIght will repeat a scene if you try to apply it to a greater number of
lights than presets in the scene. Thus, a preset is just a special case
of a scene, and subsequently that concept has been removed from the UI.
However, I still use the term internally to mean an update payload
before the "human-friendly to Hue format" scaling is applied.