Add TransitWidget
This widget displays the minutes-to-arrival of the next two trains. I'll
never have to sit outside in the rain again.

This is a great example of the kind of information that belongs in the
panel: things that are usually at least mildly interesting and which are
occasionally critically important; things that tend to creep up on you
unless you keep tabs on them.

A future enhancement: dial back the refresh frequency during periods
when I'm not likely to be using public transit.